Moved logging tests from feature to unit
